+++ apps/doc/src/app/components/table/table-example.component.ts
@@ -40,6 +40,24 @@ export class TableExampleComponent {
     LESS: import('./examples/table-basic-example/table-basic-example.component.less?raw'),
   };
 
+  public readonly exampleTreeTable: TuiDocExample = {
+    TypeScript: import('./examples/table-tree-example/table-tree-example.component?raw'),
+    HTML: import('./examples/table-tree-example/table-tree-example.component.html?raw'),
+    LESS: import('./examples/table-tree-example/table-tree-example.component.less?raw'),
+  };
+
+  public readonly exampleEmptyTable: TuiDocExample = {
+    TypeScript: import('./examples/table-empty-example/table-empty-example.component?raw'),
+    HTML: import('./examples/table-empty-example/table-empty-example.component.html?raw'),
+    LESS: import('./examples/table-empty-example/table-empty-example.component.less?raw'),
+  };
+
+  public readonly exampleLoadingTable: TuiDocExample = {
+    TypeScript: import('./examples/table-loading-example/table-loading-example.component?raw'),
+    HTML: import('./examples/table-loading-example/table-loading-example.component.html?raw'),
+    LESS: import('./examples/table-loading-example/table-loading-example.component.less?raw'),
+  };
+
   public exampleBorderStyleTable: TuiDocExample = {
     TypeScript: import('./examples/table-border-style-example/table-border-style-example.component?raw'),
     HTML: import('./examples/table-border-style-example/table-border-style-example.component.html?raw'),
@@ -93,11 +111,13 @@ export class TableExampleComponent {
     HTML: import('./examples/table-search-example/table-search-example.component.html?raw'),
     LESS: import('./examples/table-search-example/table-search-example.component.less?raw'),
   };
+
   public exampleSortTable: TuiDocExample = {
     TypeScript: import('./examples/table-sort-example/table-sort-example.component?raw'),
     HTML: import('./examples/table-sort-example/table-sort-example.component.html?raw'),
     LESS: import('./examples/table-sort-example/table-sort-example.component.less?raw'),
   };
+
   public exampleServerSortTable: TuiDocExample = {
     TypeScript: import('./examples/table-server-sort-example/table-server-sort-example.component?raw'),
     HTML: import('./examples/table-server-sort-example/table-server-sort-example.component.html?raw'),