+++ libs/components/src/lib/components/table/directives/empty.directive.ts
@@ -1,0 +1,8 @@
+import { Directive, Inject, TemplateRef } from '@angular/core';
+
+@Directive({
+  selector: `ng-template[prizmTableEmpty]`,
+})
+export class PrizmTableEmptyDirective {
+  constructor(@Inject(TemplateRef) readonly template: TemplateRef<unknown>) {}
+}